requirements.txt
annotated-types==0.7.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.18.0
pytest==8.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.2
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing-extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
